Moving the store to the Accounts namespace.
