+++ GoogleCloudExtension/GoogleCloudExtension/ResetPassword/ResetPasswordWindow.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Google.Apis.Compute.v1.Data;
-using GoogleCloudExtension.ManageWindowsCredentials;
+using GoogleCloudExtension.Accounts;
 using GoogleCloudExtension.Theming;
 using System;
